requirements.txt
PyYAML==6.0.2
yt-dlp==2024.12.13
